pyproject.toml
[project]
name = "rfid"
version = "0.1.0"
description = "Microservicio para impresión de etiquetas RFID"
authors = [
    {name = "angelvelazquez999",email = "angelvelazquezesparza@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "firebird-driver>=1.10.0",
    "pydantic>=2.0.0",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
